requirements-dev.txt
-r requirements.txt

# For tests
coverage==3.7.1
flake8-per-file-ignores==0.4.0
flake8==3.5.0
pycodestyle==2.3.0
pytest-cov==2.5.1
pytest==3.2.3
python-coveralls==2.5.0
python-dotenv==0.9.1
requests-mock==0.6.0
testfixtures==6.0.2